requirements.txt
Django==1.6.11
django-braces==1.8.0
django-oauth-toolkit==0.8.1
djangorestframework==3.1.3
oauthlib==0.7.2
PyJWT==1.3.0
python-openid==2.2.5
python-social-auth==0.2.10
requests==2.7.0
requests-oauthlib==0.5.0
six==1.9.0
South==1.0.2
